+++ src/main/java/es/in2/vcverifier/oauth2/domain/model/AuthorizationContext.java
@@ -11,6 +11,8 @@ public record AuthorizationContext(
         String originalRequestURL,
         String requestUri,
         String codeChallenge,
-        String codeChallengeMethod
+        String codeChallengeMethod,
+        String portalUrl,
+        String contextPath
 ) {
 }